ui: move screens in this package
