+++ app/src/main/java/page/ooooo/geoshare/ConversionActivity.kt
@@ -12,6 +12,7 @@ import androidx.compose.ui.platform.LocalClipboard
 import androidx.compose.ui.platform.LocalContext
 import dagger.hilt.android.AndroidEntryPoint
 import page.ooooo.geoshare.lib.ConversionRunContext
+import page.ooooo.geoshare.ui.ConversionNavigation
 import page.ooooo.geoshare.ui.theme.AppTheme
 
 @AndroidEntryPoint
