+++ app/src/main/java/page/ooooo/geoshare/MainActivity.kt
@@ -11,6 +11,7 @@ import androidx.compose.ui.platform.LocalClipboard
 import androidx.compose.ui.platform.LocalContext
 import dagger.hilt.android.AndroidEntryPoint
 import page.ooooo.geoshare.lib.ConversionRunContext
+import page.ooooo.geoshare.ui.MainNavigation
 import page.ooooo.geoshare.ui.theme.AppTheme
 
 @AndroidEntryPoint
