+++ app/src/main/java/page/ooooo/geoshare/ui/components/ScreenshotOpenByDefaultMapApp.kt
@@ -1,4 +1,4 @@
-package page.ooooo.geoshare
+package page.ooooo.geoshare.ui.components
 
 import android.content.res.Configuration
 import androidx.compose.foundation.background
@@ -13,6 +13,7 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Devices
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.IntSize
+import page.ooooo.geoshare.R
 import page.ooooo.geoshare.ui.theme.*
 
 @Composable
