+++ app/src/main/java/page/ooooo/geoshare/ui/components/UserPreferencesDetailPane.kt
@@ -1,4 +1,4 @@
-package page.ooooo.geoshare.ui
+package page.ooooo.geoshare.ui.components
 
 import android.content.res.Configuration
 import androidx.compose.foundation.layout.Arrangement
@@ -18,8 +18,7 @@ import androidx.datastore.preferences.core.MutablePreferences
 import page.ooooo.geoshare.R
 import page.ooooo.geoshare.data.di.defaultFakeUserPreferences
 import page.ooooo.geoshare.data.local.preferences.UserPreferencesValues
-import page.ooooo.geoshare.ui.components.Headline
-import page.ooooo.geoshare.ui.components.ParagraphHtml
+import page.ooooo.geoshare.ui.UserPreferencesGroup
 import page.ooooo.geoshare.ui.theme.AppTheme
 import page.ooooo.geoshare.ui.theme.LocalSpacing
