+++ app/src/main/java/page/ooooo/geoshare/ui/components/UserPreferencesListPane.kt
@@ -1,4 +1,4 @@
-package page.ooooo.geoshare.ui
+package page.ooooo.geoshare.ui.components
 
 import android.content.res.Configuration
 import androidx.compose.foundation.background
@@ -18,7 +18,8 @@ import androidx.compose.ui.tooling.preview.Preview
 import page.ooooo.geoshare.R
 import page.ooooo.geoshare.data.di.defaultFakeUserPreferences
 import page.ooooo.geoshare.data.local.preferences.UserPreferencesValues
-import page.ooooo.geoshare.ui.components.Headline
+import page.ooooo.geoshare.ui.UserPreferencesGroup
+import page.ooooo.geoshare.ui.UserPreferencesGroupId
 import page.ooooo.geoshare.ui.theme.AppTheme
 
 @OptIn(ExperimentalMaterial3Api::class)
